+++ Assets/Scripts/Weapon/BulletPattern.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public abstract class BulletPattern : ScriptableObject
 {
-    public abstract void Shoot(Vector2 m_shootPosition, Transform p_target, Bullet p_bullet, int p_damage);
+    public abstract void Shoot(Vector2 m_shootPosition, Transform p_target, Bullet p_bullet, float p_damage);
 }
