+++ Assets/Scripts/Weapon/Patterns/SingleShot.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Singleshot", menuName = "Pattern/Singleshot")]
 public class SingleShot : BulletPattern
 {
-    public override void Shoot(Vector2 m_position, Transform p_target, Bullet p_bullet, int p_damage)
+    public override void Shoot(Vector2 m_position, Transform p_target, Bullet p_bullet, float p_damage)
     {
         //TODO: Use pool
         Bullet bullet = Instantiate(p_bullet, m_position, Quaternion.identity);
